[GFC][Cleanup] Move the grid sizing algorithm into a GridSizer class https://bugs.webkit.org/show_bug.cgi?id=322534 rdar://problem/185829927
Reviewed by Tim Nguyen.

GridLayout::layout() ran the grid sizing algorithm inline, alongside grid item
placement, grid item layout, and alignment. Move this logic into its own class
so that it is a bit more modular.

GridSizer takes the grid formatting context and the layout state, and each entry
point takes the placed grid items plus the track sizing functions for both axes.
Building those track sizing functions stays in GridLayout, since grid item sizing
still needs them once the tracks have been sized, so there is nothing to be
gained from moving their construction and handing it back out again.

The fast path where the caller only needs the column sizes resolved by step 1 is
sizeColumnsOnlyFastPath(), which delegates to sizeColumnTracks(), the counterpart
to sizeRowTracks().

No change in behavior.

Canonical link: https://commits.webkit.org/319845@main
