+++ Source/WebCore/layout/formattingContexts/grid/GridLayout.cpp
@@ -31,17 +31,15 @@
 #include "GridItemRect.h"
 #include "GridLayoutState.h"
 #include "GridLayoutUtils.h"
+#include "GridSizer.h"
 #include "ImplicitGrid.h"
 #include "LayoutBoxGeometry.h"
 #include "LayoutElementBox.h"
-#include "NotImplemented.h"
 #include "PlacedGridItem.h"
 #include "StyleComputedStyle+GettersInlines.h"
-#include "TrackSizingAlgorithm.h"
 #include "TrackSizingFunctions.h"
 #include "UnplacedGridItem.h"
 #include "UsedTrackSizes.h"
-#include <wtf/Range.h>
 #include <wtf/Vector.h>
 
 namespace WebCore {
@@ -127,14 +125,14 @@ GridLayoutResult GridLayout::layout(const UnplacedGridItems& unplacedGridItems, 
     // on its block size). Steps 2-4 cannot change the column sizes, so size the columns alone and
     // skip row sizing, grid-item layout, and alignment.
     if (scope == GridLayoutScope::ColumnSizingOnly) {
-        TrackSizes columnSizes = sizeColumnTracks(placedGridItems, columnTrackSizingFunctionsList, rowTrackSizingFunctionsList, gridLayoutState);
+        TrackSizes columnSizes = GridSizer { formattingContext, gridLayoutState }.sizeColumnsOnlyFastPath(placedGridItems, columnTrackSizingFunctionsList, rowTrackSizingFunctionsList);
         return { { columnSizes, { } }, { } };
     }
 
     // 2. FIXME: Find the size of the grid container.
 
     // 3. Given the resulting grid container size, run the Grid Sizing Algorithm to size the grid.
-    UsedTrackSizes usedTrackSizes = performGridSizingAlgorithm(gridLayoutState, placedGridItems, columnTrackSizingFunctionsList, rowTrackSizingFunctionsList);
+    UsedTrackSizes usedTrackSizes = GridSizer { formattingContext, gridLayoutState }.sizeGrid(placedGridItems, columnTrackSizingFunctionsList, rowTrackSizingFunctionsList);
 
     CheckedRef formattingContextRootStyle = formattingContext.root().style();
     auto gridAreaSizes = computeGridAreaSizes(placedGridItems, gridLayoutState.usedColumnGap, gridLayoutState.usedRowGap, usedTrackSizes);
@@ -299,145 +297,6 @@ TrackSizingFunctionsList GridLayout::trackSizingFunctions(size_t totalTracksCoun
 
     ASSERT(trackSizingFunctions.size() == totalTracksCount);
     return trackSizingFunctions;
-}
-
-// If calculating the layout of a grid item in this step depends on the available space in the block axis,
-// assume the available space that it would have if any row with a definite max track sizing function
-// had that size and all other rows were infinite.
-static Vector<LayoutUnit> rowSizesForFirstIterationColumnSizing(const TrackSizingFunctionsList& rowTrackSizingFunctionsList, std::optional<LayoutUnit> gridContainerInnerInlineSize)
-{
-    return rowTrackSizingFunctionsList.map([&gridContainerInnerInlineSize](const TrackSizingFunctions& trackSizingFunctions) {
-        return WTF::switchOn(trackSizingFunctions.max,
-            [&](const Style::GridTrackBreadthLength::Fixed& fixedValue) {
-                return Style::evaluate<LayoutUnit>(fixedValue, trackSizingFunctions.zoom);
-            },
-            [&](const Style::GridTrackBreadthLength::Percentage& percentageValue) {
-                ASSERT(gridContainerInnerInlineSize, "The formatting context should have transformed this track size to auto");
-                return Style::evaluate<LayoutUnit>(percentageValue, *gridContainerInnerInlineSize);
-            },
-            [&](const Style::GridTrackBreadth::Calc calculatedValue) -> LayoutUnit {
-                ASSERT(gridContainerInnerInlineSize, "The formatting context should have transformed this track size to auto");
-                return Style::evaluate<LayoutUnit>(calculatedValue, *gridContainerInnerInlineSize, trackSizingFunctions.zoom);
-            },
-            [](const CSS::Keyword::MinContent&) -> LayoutUnit {
-                return LayoutUnit::max();
-            },
-            [](const CSS::Keyword::MaxContent&) {
-                return LayoutUnit::max();
-            },
-            [](const CSS::Keyword::Auto&) -> LayoutUnit {
-                return LayoutUnit::max();
-            },
-            [](const Style::GridTrackBreadth::Flex&) -> LayoutUnit {
-                return LayoutUnit::max();
-            },
-            [](const auto&) -> LayoutUnit {
-                ASSERT_NOT_IMPLEMENTED_YET();
-                return { };
-            });
-    });
-}
-
-// During track sizing we may need to get different types of size contributions for a grid item.
-// Getting a contribution in a specific dimension may require knowing the available space in
-// the opposite dimension. For each of these cases, the spec defines how to compute the available space.
-static LayoutUnit NODELETE oppositeAxisConstraintForTrackSizing(const Vector<LayoutUnit>& oppositeAxisTrackSizes, const WTF::Range<size_t> oppositeAxisSpan)
-{
-    auto totalAvailableSpaceFromSpannedTracks = 0_lu;
-    for (auto oppositeAxisLineIndex : std::views::iota(oppositeAxisSpan.begin(), oppositeAxisSpan.end())) {
-        auto& oppositeAxisTrackSize = oppositeAxisTrackSizes[oppositeAxisLineIndex];
-        if (oppositeAxisTrackSize == LayoutUnit::max())
-            return oppositeAxisTrackSize;
-
-        totalAvailableSpaceFromSpannedTracks += oppositeAxisTrackSize;
-    }
-    return totalAvailableSpaceFromSpannedTracks;
-}
-
-// 1. https://www.w3.org/TR/css-grid-1/#algo-grid-sizing — step 1.
-// First, the track sizing algorithm is used to resolve the sizes of the grid columns.
-// If calculating the layout of a grid item in this step depends on the available space in the block axis,
-// assume the available space that it would have if any row with a definite max track sizing function had
-// that size and all other rows were infinite. If both the grid container and all tracks have definite sizes,
-// also apply align-content to find the final effective size of any gaps spanned by such items; otherwise
-// ignore the effects of track alignment in this estimation.
-TrackSizes GridLayout::sizeColumnTracks(const PlacedGridItems& placedGridItems, const TrackSizingFunctionsList& columnTrackSizingFunctionsList,
-    const TrackSizingFunctionsList& rowTrackSizingFunctionsList, const GridLayoutState& layoutState) const
-{
-    auto& layoutConstraints = layoutState.gridLayoutConstraints;
-
-    auto columnFreeSpaceScenario = layoutConstraints.inlineAxis.scenario();
-    std::optional<LayoutUnit> inlineAxisAvailableSpace = columnFreeSpaceScenario == AxisConstraint::FreeSpaceScenario::Definite
-        ? std::optional(layoutConstraints.inlineAxis.availableSpace())
-        : std::nullopt;
-    auto rowSizesForFirstColumnSizing = rowSizesForFirstIterationColumnSizing(rowTrackSizingFunctionsList, inlineAxisAvailableSpace);
-
-    auto columnTrackSizingItems = placedGridItems.map([&](const PlacedGridItem& gridItem) -> TrackSizingItem {
-        auto rowSpan = WTF::Range<size_t> { gridItem.rowStartLine(), gridItem.rowEndLine() };
-        // The inline grid area is indefinite while sizing columns, so the item's cyclic percentage padding resolves against zero.
-        auto usedInlineBorderAndPadding = formattingContext().integrationUtils().borderAndPaddingForGridItem(gridItem.layoutBox(), 0_lu).first;
-        return { gridItem, gridItem.inlineAxisSizes(), usedInlineBorderAndPadding,
-            { gridItem.columnStartLine(), gridItem.columnEndLine() }, oppositeAxisConstraintForTrackSizing(rowSizesForFirstColumnSizing, rowSpan) };
-    });
-
-    return TrackSizingAlgorithm::sizeTracks(columnTrackSizingItems, columnTrackSizingFunctionsList,
-        layoutConstraints.inlineAxis, GridItemSizingFunctions::inlineAxis(formattingContext().integrationUtils()),
-        layoutState.usedColumnGap, layoutState.usedJustifyContent);
-}
-
-// 2. https://www.w3.org/TR/css-grid-1/#algo-grid-sizing — step 2.
-// Next, the track sizing algorithm resolves the sizes of the grid rows.
-// To find the inline-axis available space for any items whose block-axis size contributions
-// require it, use the grid column sizes calculated in the previous step.
-TrackSizes GridLayout::sizeRowTracks(const PlacedGridItems& placedGridItems, const TrackSizes& columnSizes,
-    const TrackSizingFunctionsList& rowTrackSizingFunctionsList, const GridLayoutState& layoutState) const
-{
-    auto& layoutConstraints = layoutState.gridLayoutConstraints;
-
-    auto rowTrackSizingItems = placedGridItems.map([&](const PlacedGridItem& gridItem) -> TrackSizingItem {
-        auto columnSpan = WTF::Range<size_t> { gridItem.columnStartLine(), gridItem.columnEndLine() };
-        auto columnConstraint = oppositeAxisConstraintForTrackSizing(columnSizes, columnSpan);
-        auto gridAreaInlineSize = GridLayoutUtils::gridAreaDimensionSize(gridItem.columnStartLine(), gridItem.columnEndLine(), columnSizes, layoutState.usedColumnGap);
-        auto usedBlockBorderAndPadding = formattingContext().integrationUtils().borderAndPaddingForGridItem(gridItem.layoutBox(), gridAreaInlineSize).second;
-        return { gridItem, gridItem.blockAxisSizes(), usedBlockBorderAndPadding,
-            { gridItem.rowStartLine(), gridItem.rowEndLine() }, columnConstraint };
-    });
-
-    return TrackSizingAlgorithm::sizeTracks(rowTrackSizingItems, rowTrackSizingFunctionsList,
-        layoutConstraints.blockAxis, GridItemSizingFunctions::blockAxis(formattingContext()),
-        layoutState.usedRowGap, layoutState.usedAlignContent);
-}
-
-// https://www.w3.org/TR/css-grid-1/#algo-grid-sizing
-UsedTrackSizes GridLayout::performGridSizingAlgorithm(const GridLayoutState& layoutState, const PlacedGridItems& placedGridItems,
-    const TrackSizingFunctionsList& columnTrackSizingFunctionsList, const TrackSizingFunctionsList& rowTrackSizingFunctionsList) const
-{
-    // 1. First, the track sizing algorithm is used to resolve the sizes of the grid columns.
-    // If both the grid container and all tracks have definite sizes, also apply align-content
-    // to find the final effective size of any gaps spanned by such items; otherwise ignore
-    // the effects of track alignment in this estimation.
-    auto columnSizes = sizeColumnTracks(placedGridItems, columnTrackSizingFunctionsList, rowTrackSizingFunctionsList, layoutState);
-
-    // 2. Next, the track sizing algorithm resolves the sizes of the grid rows.
-    auto rowSizes = sizeRowTracks(placedGridItems, columnSizes, rowTrackSizingFunctionsList, layoutState);
-
-    // 3. Then, if the min-content contribution of any grid item has changed based on the
-    // row sizes and alignment calculated in step 2, re-resolve the sizes of the grid
-    // columns with the new min-content and max-content contributions (once only).
-    auto resolveGridColumnSizesIfAnyMinContentContributionChanged = [] {
-        notImplemented();
-    };
-    UNUSED_VARIABLE(resolveGridColumnSizesIfAnyMinContentContributionChanged);
-
-    // 4. Next, if the min-content contribution of any grid item has changed based on the
-    // column sizes and alignment calculated in step 3, re-resolve the sizes of the grid
-    // rows with the new min-content and max-content contributions (once only).
-    auto resolveGridRowSizesIfAnyMinContentContributionChanged = [] {
-        notImplemented();
-    };
-    UNUSED_VARIABLE(resolveGridRowSizesIfAnyMinContentContributionChanged);
-
-    return { columnSizes, rowSizes };
 }
 
 // https://drafts.csswg.org/css-grid-1/#auto-margins
